+++ tests/Unit/Core/DynImgGeneratorTest.php
@@ -84,13 +84,19 @@ final class DynImgGeneratorTest extends \OxidTestCase
         $this->assertEquals(getShopBasePath() . "/test1/test2/test3/12_12_12/test.jpg", $oGen->getImageTarget());
     }
 
-    public function testGetNopicImageTarget(): void
+    public function testGetNopicImageTarget(string $filename = 'nopic.jpg'): void
     {
         $oGen = $this->getMock(\OxidEsales\Eshop\Core\DynamicImageGenerator::class, array("getImageUri", "getImageName"));
         $oGen->method('getImageUri')->will($this->returnValue("/test1/test2/test3/12_12_12/test.jpg"));
         $oGen->method('getImageName')->will($this->returnValue("test.jpg"));
 
-        $this->assertEquals(getShopBasePath() . "/test1/test2/test3/12_12_12/nopic.jpg", $oGen->getNopicImageTarget());
+        $this->assertEquals(getShopBasePath() . "/test1/test2/test3/12_12_12/$filename", $oGen->getNopicImageTarget());
+    }
+
+    public function testGetNopicWebpImageTarget(): void
+    {
+        $this->setConfigParam('blConvertImagesToWebP', 1);
+        $this->testGetNopicImageTarget('nopic.webp');
     }
 
     public function testIsTargetPathValid(): void